leave cache in main cli file
