+++ src/subliminal/cli/commands/__init__.py
@@ -1,6 +1,5 @@
 """CLI subcommands."""
 
-from .cache import cache, cache_file
 from .download_best import download
 
-__all__ = ['cache', 'cache_file', 'download']
+__all__ = ['download']
